refactors negation detection
